+++ mbo/container/internal/limited_ordered.h
@@ -1,4 +1,4 @@
-// Copyright 2023 M. Boerger (helly25.com)
+// Copyright M. Boerger (helly25.com)
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -16,7 +16,7 @@
 #define MBO_CONTAINER_INTERNAL_LIMITED_ORDERED_H_
 
 #include <algorithm>
-#include <compare>  // IWYU pragma: keep
+#include <compare>   // IWYU pragma: keep
 #include <concepts>  // IWYU pragma: keep
 #include <initializer_list>
 #include <memory>
@@ -29,20 +29,34 @@
 #include "mbo/types/compare.h"              // IWYU pragma: export
 #include "mbo/types/traits.h"
 
+// If `mbo/container/internal/limited_ordered_config.h` is available, then include that. In order to
+// make indexers work, we also include the generator "...in" as a fallback.
+#if __has_include("mbo/container/internal/limited_ordered_config.h")
+# include "mbo/container/internal/limited_ordered_config.h"
+#else
+# include "mbo/container/internal/limited_ordered_config.h.in"
+#endif
+
 #ifdef MBO_FORCE_INLINE
-#undef MBO_FORCE_INLINE
+# undef MBO_FORCE_INLINE
 #endif
 #define MBO_FORCE_INLINE
+
 // __attribute__((always_inline))
+
+#ifdef MBO_ALWAYS_INLINE
+# undef MBO_ALWAYS_INLINE
+#endif
+#define MBO_ALWAYS_INLINE __attribute__((always_inline))
 
 namespace mbo::container::container_internal {
 
 #ifdef LV_REQUIRE
-#undef LV_REQUIRE
+# undef LV_REQUIRE
 #endif
 
 // NOLINTNEXTLINE(cppcoreguidelines-macro-usage)
-#define LV_REQUIRE(severity, condition) \
+#define LV_REQUIRE(severity, condition)                      \
   /* NOLINTNEXTLINE(bugprone-switch-missing-default-case) */ \
   ABSL_LOG_IF(severity, !(condition))
 
@@ -59,14 +73,31 @@ concept LimitedOrderedValid =  //
     std::move_constructible<std::remove_const_t<Key>> && std::move_constructible<Mapped>
     && LimitedOrderedValidImpl<Key, Mapped, Value>;
 
-template<typename Key, typename Mapped, typename Value, auto options, typename Compare = types::CompareLess<Key>>
+template<typename Key, typename Mapped, typename Value, auto options, typename Compare = std::less<Key>>
 requires(LimitedOrderedValid<Key, Mapped, Value>)
 class LimitedOrdered {
  protected:
+  static constexpr bool kKeyOnly = std::same_as<Key, Value>;  // true = set, false = map (of pairs).
+
+  // Size and Options management.
   static_assert(IsLimitedOptionsOrSize<decltype(options)>);
   using Options = decltype(MakeLimitedOptions<options>());
   static_assert(std::is_trivially_destructible_v<Value> || !Options::Has(LimitedOptionsFlag::kEmptyDestructor));
   static constexpr std::size_t Capacity = Options::kCapacity;
+
+  static constexpr bool kOptimizeIndexOf = !Options::Has(LimitedOptionsFlag::kNoOptimizeIndexOf);
+  static constexpr bool kCustomIndexOfBeyondUnroll = Options::Has(LimitedOptionsFlag::kCustomIndexOfBeyondUnroll);
+
+  static constexpr std::size_t kUnrollMaxCapacityLimit = 32;                    // The maximum supported in code.
+  static constexpr std::size_t kUnrollMaxCapacity = kUnrollMaxCapacityDefault;  // MUST MATCH `index_of`.
+  static_assert(
+      kUnrollMaxCapacity >= 4 && kUnrollMaxCapacity <= kUnrollMaxCapacityLimit,
+      "Check documentation for `mbo::container::container_internal::kUnrollMaxCapacityDefault`.");
+
+  // Must declare each other as friends so that we can correctly move from other.
+  template<typename OK, typename OM, typename OV, auto OtherN, typename Comp>
+  requires(LimitedOrderedValid<OK, OM, OV>)
+  friend class LimitedOrdered;
 
   struct None final {};
 
@@ -89,14 +120,6 @@ class LimitedOrdered {
     None none;
   };
 
-  // Must declare each other as friends so that we can correctly move from other.
-  template<typename OK, typename OM, typename OV, auto OtherN, typename Comp>
-  requires(LimitedOrderedValid<OK, OM, OV>)
-  friend class LimitedOrdered;
-
-  static constexpr bool kKeyOnly = std::same_as<Key, Value>;  // true = set, false = map (of pairs).
-  static constexpr std::size_t kUnrollMaxCapacity = 8;        // MUST MATCH `index_of`.
-
  public:
   using key_type = Key;
   using value_type = Value;
@@ -111,15 +134,21 @@ class LimitedOrdered {
   static constexpr size_type npos = static_cast<size_type>(-1);  // Result of `index_of` if not found.
 
   struct ValueCompare {
-    static constexpr const Key& GetKey(const Key& key) noexcept { return key; }
+    MBO_ALWAYS_INLINE static constexpr const Key& GetKey(const Key& key) noexcept { return key; }
 
-    static constexpr const Key& GetKey(const value_type& val) noexcept
+    MBO_ALWAYS_INLINE static constexpr const Key& GetKey(const value_type& val) noexcept
     requires(!kKeyOnly)
     {
       return val.first;
     }
 
-    static constexpr const Key& GetKey(const Data& data) noexcept { return GetKey(data.data); }
+    MBO_ALWAYS_INLINE static constexpr const Key& GetKey(const Data& data) noexcept {
+      if constexpr (kKeyOnly) {
+        return data.data;
+      } else {
+        return data.data.first;
+      }
+    }
 
     constexpr ~ValueCompare() noexcept = default;
 
@@ -142,13 +171,15 @@ class LimitedOrdered {
        std::same_as<std::remove_cvref_t<R>, Value> ||
        std::same_as<std::remove_cvref_t<R>, Data>)
     )  // clang-format on
-    constexpr bool operator()(const L& lhs, const R& rhs) const noexcept {
+    MBO_FORCE_INLINE constexpr bool operator()(const L& lhs, const R& rhs) const noexcept {
       return key_comp(GetKey(lhs), GetKey(rhs));
     }
 
     const key_compare key_comp;
   };
 
+  // NOTE: The name is misleading but must adhere to the STL: The comparator only
+  // compares the `first` part (the key) of mapped values.
   using value_compare = std::conditional_t<kKeyOnly, Compare, ValueCompare>;
 
   class const_iterator {
@@ -162,13 +193,13 @@ class LimitedOrdered {
 
     constexpr const_iterator() noexcept : pos_(nullptr) {}  // Needed for STL
 
-    constexpr explicit const_iterator(const Data* pos) noexcept : pos_(pos) {}
+    MBO_ALWAYS_INLINE constexpr explicit const_iterator(const Data* pos) noexcept : pos_(pos) {}
 
-    constexpr explicit operator reference() const noexcept { return pos_->data; }
+    MBO_ALWAYS_INLINE constexpr explicit operator reference() const noexcept { return pos_->data; }
 
-    constexpr reference operator*() const noexcept { return pos_->data; }
+    MBO_ALWAYS_INLINE constexpr reference operator*() const noexcept { return pos_->data; }
 
-    constexpr pointer operator->() const noexcept { return &pos_->data; }
+    MBO_ALWAYS_INLINE constexpr pointer operator->() const noexcept { return &pos_->data; }
 
     constexpr const_iterator& operator++() noexcept {
       ++pos_;
@@ -230,11 +261,17 @@ class LimitedOrdered {
       return const_iterator(lhs + rhs.pos_);
     }
 
-    friend constexpr auto operator<=>(const_iterator lhs, const_iterator rhs) noexcept { return lhs.pos_ <=> rhs.pos_; }
+    MBO_ALWAYS_INLINE friend constexpr auto operator<=>(const_iterator lhs, const_iterator rhs) noexcept {
+      return lhs.pos_ <=> rhs.pos_;
+    }
 
-    friend constexpr bool operator<(const_iterator lhs, const_iterator rhs) noexcept { return lhs.pos_ < rhs.pos_; }
+    MBO_ALWAYS_INLINE friend constexpr bool operator<(const_iterator lhs, const_iterator rhs) noexcept {
+      return lhs.pos_ < rhs.pos_;
+    }
 
-    friend constexpr bool operator==(const_iterator lhs, const_iterator rhs) noexcept { return lhs.pos_ == rhs.pos_; }
+    MBO_ALWAYS_INLINE friend constexpr bool operator==(const_iterator lhs, const_iterator rhs) noexcept {
+      return lhs.pos_ == rhs.pos_;
+    }
 
    private:
     const Data* pos_;
@@ -251,15 +288,15 @@ class LimitedOrdered {
 
     constexpr iterator() noexcept : pos_(nullptr) {}  // Needed for STL
 
-    constexpr explicit iterator(Data* pos) noexcept : pos_(pos) {}
+    MBO_ALWAYS_INLINE constexpr explicit iterator(Data* pos) noexcept : pos_(pos) {}
 
-    constexpr explicit operator reference() const noexcept { return pos_->data; }
+    MBO_ALWAYS_INLINE constexpr explicit operator reference() const noexcept { return pos_->data; }
 
-    constexpr reference operator*() const noexcept { return pos_->data; }
+    MBO_ALWAYS_INLINE constexpr reference operator*() const noexcept { return pos_->data; }
 
-    constexpr pointer operator->() const noexcept { return &pos_->data; }
+    MBO_ALWAYS_INLINE constexpr pointer operator->() const noexcept { return &pos_->data; }
 
-    constexpr explicit operator const_iterator() const noexcept { return const_iterator(pos_); }
+    MBO_ALWAYS_INLINE constexpr explicit operator const_iterator() const noexcept { return const_iterator(pos_); }
 
     constexpr iterator& operator++() noexcept {
       ++pos_;
@@ -309,11 +346,17 @@ class LimitedOrdered {
 
     friend constexpr iterator operator+(difference_type lhs, iterator rhs) noexcept { return iterator(lhs + rhs.pos_); }
 
-    friend constexpr auto operator<=>(iterator lhs, iterator rhs) noexcept { return lhs.pos_ <=> rhs.pos_; }
+    MBO_ALWAYS_INLINE friend constexpr auto operator<=>(iterator lhs, iterator rhs) noexcept {
+      return lhs.pos_ <=> rhs.pos_;
+    }
 
-    friend constexpr bool operator<(iterator lhs, iterator rhs) noexcept { return lhs.pos_ < rhs.pos_; }
+    MBO_ALWAYS_INLINE friend constexpr bool operator<(iterator lhs, iterator rhs) noexcept {
+      return lhs.pos_ < rhs.pos_;
+    }
 
-    friend constexpr bool operator==(iterator lhs, iterator rhs) noexcept { return lhs.pos_ == rhs.pos_; }
+    MBO_ALWAYS_INLINE friend constexpr bool operator==(iterator lhs, iterator rhs) noexcept {
+      return lhs.pos_ == rhs.pos_;
+    }
 
    private:
     Data* pos_;
@@ -444,68 +487,69 @@ class LimitedOrdered {
     return *this;
   }
 
-  template<typename OK, typename OM, typename OV, auto OtherN, typename OtherCompare>
-  requires(
-      std::convertible_to<OK, Key> && std::convertible_to<OM, Mapped>
-      && MakeLimitedOptions<OtherN>().kCapacity <= Capacity)
-  constexpr explicit LimitedOrdered(LimitedOrdered<OK, OM, OV, OtherN, OtherCompare>&& other) noexcept {
-    for (auto it = other.begin(); it < other.end(); ++it) {
-      if constexpr (kKeyOnly) {
-        emplace(*it);
-      } else {
-        emplace(it->first, it->second);
-      }
-    }
-    other.size_ = 0;
-  }
-
-  template<typename OK, typename OM, typename OV, auto OtherN, typename OtherCompare>
-  requires(
-      std::convertible_to<OK, Key> && std::convertible_to<OM, Mapped>
-      && MakeLimitedOptions<OtherN>().kCapacity <= Capacity)
-  constexpr LimitedOrdered& operator=(LimitedOrdered<OK, OM, OV, OtherN, OtherCompare>&& other) noexcept {
-    clear();
-    for (auto it = other.begin(); it < other.end(); ++it) {
-      if constexpr (kKeyOnly) {
-        emplace(*it);
-      } else {
-        emplace(it->first, it->second);
-      }
-    }
-    other.size_ = 0;
-    return *this;
-  }
-
   // Find and search: lower_bound, upper_bound, equal_range, find, contains, count
 
-  constexpr iterator lower_bound(const Key& key) { return std::lower_bound(begin(), end(), key, val_comp_); }
-
-  constexpr const_iterator lower_bound(const Key& key) const {
+  MBO_FORCE_INLINE constexpr iterator lower_bound(const Key& key) {
     return std::lower_bound(begin(), end(), key, val_comp_);
   }
 
-  constexpr iterator upper_bound(const Key& key) { return std::upper_bound(begin(), end(), key, val_comp_); }
+  MBO_FORCE_INLINE constexpr const_iterator lower_bound(const Key& key) const {
+    return std::lower_bound(begin(), end(), key, val_comp_);
+  }
 
-  constexpr const_iterator upper_bound(const Key& key) const {
+  MBO_FORCE_INLINE constexpr iterator upper_bound(const Key& key) {
     return std::upper_bound(begin(), end(), key, val_comp_);
   }
 
-  // NOLINTBEGIN(*-magic-numbers,cppcoreguidelines-macro-usage,readability-function-cognitive-complexity)
-  MBO_FORCE_INLINE constexpr std::size_t index_of(const Key& key) const
-  requires(types::IsCompareLess<Compare> && Capacity <= kUnrollMaxCapacity)
+  MBO_FORCE_INLINE constexpr const_iterator upper_bound(const Key& key) const {
+    return std::upper_bound(begin(), end(), key, val_comp_);
+  }
+
+  // NOLINTBEGIN(*-magic-numbers,*-macro-usage,*-function-size,readability-function-cognitive-complexity)
+  MBO_ALWAYS_INLINE constexpr std::size_t index_of(const Key& key) const
+  requires(kOptimizeIndexOf && Capacity <= kUnrollMaxCapacity)
   {
-#define MBO_CASE_LIMITED_POS_COMP(POS)                              \
-  case POS:                                                         \
-    if constexpr ((POS) >= Capacity) {                              \
-      return npos;                                                  \
-    } else {                                                        \
-      if (key_comp_.Compare(key, GetKey(values_[POS].data)) == 0) { \
-        return POS;                                                 \
-      }                                                             \
-    }                                                               \
-    do /* NOLINT(cppcoreguidelines-avoid-do-while) */ {             \
-    } while (0)
-    switch (size_ - 1) {
+#define MBO_CASE_LIMITED_POS_COMP(POS)                                     \
+  static_assert((POS) + 1 <= kUnrollMaxCapacityLimit);                     \
+  case ((POS) + 1):                                                        \
+    if constexpr ((POS) >= Capacity) {                                     \
+      return npos;                                                         \
+    } else if constexpr (mbo::types::IsCompareLess<Compare>) {             \
+      const auto comp = key_comp_.Compare(key, GetKey(values_[POS].data)); \
+      if (comp >= 0) [[unlikely]] {                                        \
+        if (comp > 0) [[likely]] {                                         \
+          return npos;                                                     \
+        } else [[unlikely]] {                                              \
+          return (POS);                                                    \
+        }                                                                  \
+      }                                                                    \
+    } else {                                                               \
+      if (!key_comp_(key, GetKey(values_[POS].data))) [[unlikely]] {       \
+        if (key_comp_(GetKey(values_[POS].data), key)) [[likely]] {        \
+          return npos;                                                     \
+        } else [[unlikely]] {                                              \
+          return POS;                                                      \
+        }                                                                  \
+      }                                                                    \
+    }                                                                      \
+    [[fallthrough]]
+    switch (size_) {
+      MBO_CASE_LIMITED_POS_COMP(31);
+      MBO_CASE_LIMITED_POS_COMP(30);
+      MBO_CASE_LIMITED_POS_COMP(29);
+      MBO_CASE_LIMITED_POS_COMP(28);
+      MBO_CASE_LIMITED_POS_COMP(27);
+      MBO_CASE_LIMITED_POS_COMP(26);
+      MBO_CASE_LIMITED_POS_COMP(25);
+      MBO_CASE_LIMITED_POS_COMP(24);
+      MBO_CASE_LIMITED_POS_COMP(23);
+      MBO_CASE_LIMITED_POS_COMP(22);
+      MBO_CASE_LIMITED_POS_COMP(21);
+      MBO_CASE_LIMITED_POS_COMP(20);
+      MBO_CASE_LIMITED_POS_COMP(19);
+      MBO_CASE_LIMITED_POS_COMP(18);
+      MBO_CASE_LIMITED_POS_COMP(17);
+      MBO_CASE_LIMITED_POS_COMP(16);
       MBO_CASE_LIMITED_POS_COMP(15);
       MBO_CASE_LIMITED_POS_COMP(14);
       MBO_CASE_LIMITED_POS_COMP(13);
@@ -522,37 +566,69 @@ class LimitedOrdered {
       MBO_CASE_LIMITED_POS_COMP(2);
       MBO_CASE_LIMITED_POS_COMP(1);
       MBO_CASE_LIMITED_POS_COMP(0);
-      default: break;  // Handles `size` - 1 == `npos`.
+      default: break;  // Handles `size_ == 0`.
     }
 #undef MBO_CASE_LIMITED_POS_COMP
     return npos;
-  }  // NOLINTEND(*-magic-numbers,cppcoreguidelines-macro-usage,readability-function-cognitive-complexity)
+  }  // NOLINTEND(*-magic-numbers,*-macro-usage,*-function-size,readability-function-cognitive-complexity)
 
-  MBO_FORCE_INLINE constexpr std::size_t index_of(const Key& key) const
-  requires(types::IsCompareLess<Compare> && Capacity > kUnrollMaxCapacity)
+  MBO_ALWAYS_INLINE constexpr std::size_t index_of(const Key& key) const
+  requires(
+      kOptimizeIndexOf && kCustomIndexOfBeyondUnroll
+      && mbo::types::IsCompareLess<Compare> && Capacity > kUnrollMaxCapacity)
   {
     std::size_t left = 0;
     std::size_t right = size_;
-    while (left < right) {
-      std::size_t pos = left + ((right - left) >> 1U);
+    while (left < right) [[likely]] {
+      const std::size_t pos = left + ((right - left) >> 1U);
       auto cmp = key_comp_.Compare(key, GetKey(values_[pos]));
-      if (cmp == 0) {
-        return pos;
-      }
-      if (cmp < 0) {
+      if (cmp < 0) [[unlikely]] {
         right = pos;
-      } else {
+      } else if (cmp > 0) {
         left = pos + 1;
+      } else {
+        return pos;  // == 0
       }
     }
     return npos;
   }
 
-  MBO_FORCE_INLINE constexpr std::size_t index_of(const Key& key) const
-  requires(!types::IsCompareLess<Compare>)
+  MBO_ALWAYS_INLINE std::size_t index_of(const Key& key) const
+  requires(
+      kOptimizeIndexOf && kCustomIndexOfBeyondUnroll
+      && !mbo::types::IsCompareLess<Compare> && Capacity > kUnrollMaxCapacity)
+  {
+    if (size_ == 0) {
+      return npos;
+    }
+    std::size_t left = 0;
+    std::size_t right = size_;
+    while (true) {
+      const std::size_t diff = (right - left) >> 1U;
+      const std::size_t pos = left + diff;
+      if (key_comp_(key, GetKey(values_[pos]))) [[likely]] {
+        if (diff == 0) [[unlikely]] {
+          return npos;
+        }
+        right = pos;
+      } else {
+        if (diff == 0) [[unlikely]] {
+          if (key_comp_(GetKey(values_[left]), key)) [[likely]] {
+            return npos;
+          } else [[unlikely]] {
+            return left;
+          }
+        }
+        left = pos;
+      }
+    }
+  }
+
+  MBO_ALWAYS_INLINE constexpr std::size_t index_of(const Key& key) const
+  requires(!kOptimizeIndexOf || (kOptimizeIndexOf && !kCustomIndexOfBeyondUnroll && Capacity > kUnrollMaxCapacity))
   {
     const_iterator it = lower_bound(key);
-    return it == end() || key_comp_(GetKey(*it), key) || key_comp_(key, GetKey(*it)) ? npos : it - begin();
+    return it == end() || key_comp_(key, GetKey(*it)) ? npos : it - begin();
   }
 
   MBO_FORCE_INLINE constexpr value_type& at_index(size_type pos) {
@@ -570,27 +646,27 @@ class LimitedOrdered {
   }
 
   MBO_FORCE_INLINE constexpr iterator find(const Key& key) {
-    if constexpr (types::IsCompareLess<Compare>) {
+    if constexpr (kOptimizeIndexOf) {
       std::size_t pos = index_of(key);
       return pos == npos ? end() : iterator(&values_[pos]);
-    } else {  // Not IsCompareLess
+    } else {  // Not kOptimizeIndexOf
       iterator it = lower_bound(key);
-      return it == end() || key_comp_(GetKey(*it), key) || key_comp_(key, GetKey(*it)) ? end() : it;
+      return it == end() || key_comp_(key, GetKey(*it)) ? end() : it;
     }
   }
 
   MBO_FORCE_INLINE constexpr const_iterator find(const Key& key) const {
-    if constexpr (types::IsCompareLess<Compare>) {
+    if constexpr (kOptimizeIndexOf) {
       std::size_t pos = index_of(key);
       return pos == npos ? end() : const_iterator(&values_[pos]);
-    } else {  // Not IsCompareLess
+    } else {  // Not kOptimizeIndexOf
       const_iterator it = lower_bound(key);
-      return it == end() || key_comp_(GetKey(*it), key) || key_comp_(key, GetKey(*it)) ? end() : it;
+      return it == end() || key_comp_(key, GetKey(*it)) ? end() : it;
     }
   }
 
   MBO_FORCE_INLINE constexpr bool contains(const Key& key) const {
-    if constexpr (types::IsCompareLess<Compare>) {
+    if constexpr (kOptimizeIndexOf) {
       return index_of(key) != npos;
     } else {
       return std::binary_search(begin(), end(), key, val_comp_);
@@ -695,7 +771,7 @@ class LimitedOrdered {
 
   template<typename... Args>
   constexpr std::pair<iterator, bool> emplace(Args&&... args) noexcept {
-    Value new_val(args...);
+    Value new_val(std::forward<Args>(args)...);
     const iterator dst = lower_bound(GetKey(new_val));
     if (dst != end() && !key_comp_(GetKey(*dst), GetKey(new_val)) && !key_comp_(GetKey(new_val), GetKey(*dst))) {
       return std::make_pair(dst, false);
@@ -756,7 +832,7 @@ class LimitedOrdered {
   constexpr size_type erase(K&& key) {
     size_type count = 0;
     while (true) {
-      auto pos = find(key);
+      auto pos = find(std::forward<K>(key));
       if (pos == end()) {
         break;
       }
@@ -884,9 +960,13 @@ class LimitedOrdered {
 
   MBO_FORCE_INLINE constexpr reverse_iterator rend() noexcept { return std::make_reverse_iterator(begin()); }
 
-  MBO_FORCE_INLINE constexpr const_reverse_iterator rend() const noexcept { return std::make_reverse_iterator(begin()); }
+  MBO_FORCE_INLINE constexpr const_reverse_iterator rend() const noexcept {
+    return std::make_reverse_iterator(begin());
+  }
 
-  MBO_FORCE_INLINE constexpr const_reverse_iterator crend() const noexcept { return std::make_reverse_iterator(cbegin()); }
+  MBO_FORCE_INLINE constexpr const_reverse_iterator crend() const noexcept {
+    return std::make_reverse_iterator(cbegin());
+  }
 
   constexpr const_pointer data() const noexcept { return &values_[0].data; }
 
@@ -911,7 +991,7 @@ class LimitedOrdered {
     return val.first;
   }
 
-  MBO_FORCE_INLINE static constexpr const Key& GetKey(const Data& data) noexcept {
+  MBO_ALWAYS_INLINE static constexpr const Key& GetKey(const Data& data) noexcept {
     if constexpr (kKeyOnly) {
       return data.data;
     } else {
@@ -937,7 +1017,11 @@ class LimitedOrdered {
 }  // namespace mbo::container::container_internal
 
 #ifdef MBO_FORCE_INLINE
-#undef MBO_FORCE_INLINE
+# undef MBO_FORCE_INLINE
+#endif
+
+#ifdef MBO_ALWAYS_INLINE
+# undef MBO_ALWAYS_INLINE
 #endif
 
 #endif  // MBO_CONTAINER_INTERNAL_LIMITED_ORDERED_H_