+++ mbo/container/limited_set_test.cc
@@ -14,15 +14,16 @@
 
 #include "mbo/container/limited_set.h"
 
-#include <iostream>
-#include <ranges>
+#include <ranges>  // IWYU pragma: keep
 #include <string>
 #include <string_view>
-#include <type_traits>
+#include <type_traits>  // IWYU pragma: keep
+#include <utility>
 
 #include "absl/log/initialize.h"
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
+#include "mbo/container/limited_options.h"
 #include "mbo/testing/matchers.h"
 
 // Clang has issues with exception tracing in ASAN, so corresponding tests must
@@ -129,6 +130,7 @@ TEST_F(LimitedSetTest, MakeInitArgBasics) {
   EXPECT_THAT(test, SizeIs(3));
   EXPECT_THAT(test, CapacityIs(7));
   EXPECT_THAT(test, ElementsAre(1, 3, 5));
+  EXPECT_THAT(test.end() - test.begin(), 3);
   EXPECT_THAT(test.find(1), test.begin());
   EXPECT_THAT(test.find(1) - test.begin(), 0);
   EXPECT_THAT(test.find(3), test.begin() + 1);
@@ -138,9 +140,11 @@ TEST_F(LimitedSetTest, MakeInitArgBasics) {
   EXPECT_THAT(test.find(5) - test.begin(), 2);
   EXPECT_THAT(test.find(0), test.end());
   EXPECT_THAT(test.emplace(0), Pair(test.begin(), true));
+  EXPECT_THAT(test.end() - test.begin(), 4);
   EXPECT_THAT(test, ElementsAre(0, 1, 3, 5));
   EXPECT_THAT(test.find(2), test.end());
   EXPECT_THAT(test.emplace(2), Pair(test.begin() + 2, true));
+  EXPECT_THAT(test.end() - test.begin(), 5);
   EXPECT_THAT(test, ElementsAre(0, 1, 2, 3, 5));
   EXPECT_THAT(test.find(6), test.end());
   EXPECT_THAT(test.emplace(6), Pair(test.end(), true));
@@ -237,7 +241,7 @@ TEST_F(LimitedSetTest, ConstructAssignFromSmaller) {
     LimitedSet<unsigned, 3> source({0U, 1U, 2U});
     LimitedSet<int, 5> target;
     ASSERT_THAT(target, IsEmpty());
-    target = std::move(source);
+    target = source;
     EXPECT_THAT(target, ElementsAre(0, 1, 2));
   }
 }
@@ -266,8 +270,8 @@ TEST_F(LimitedSetTest, ToLimitedSetStringCopy) {
 
 TEST_F(LimitedSetTest, ToLimitedSetStringMove) {
   // NOLINTBEGIN(*-avoid-c-arrays)
-  std::string array[4] = {{"0"}, {"1"}, {"2"}, {"3"}};
-  auto test = ToLimitedSet(std::move(array));
+  const std::string array[4] = {{"0"}, {"1"}, {"2"}, {"3"}};
+  auto test = ToLimitedSet(array);
   EXPECT_THAT(test, Not(IsEmpty()));
   EXPECT_THAT(test, SizeIs(4));
   EXPECT_THAT(test, CapacityIs(4));
@@ -444,48 +448,70 @@ TEST_F(LimitedSetTest, CompareDifferentType) {
   EXPECT_THAT(k42v65, Ge(k42));
 }
 
-template <std::size_t Size>
-constexpr void CompareAllTheSizes() {  // NOLINT(readability-function-cognitive-complexity)
-  LimitedSet<int, Size> data;
+template<std::size_t Size, template<typename> typename Compare, LimitedOptionsFlag... Flags>
+void CompareAllTheSizesFor() {  // NOLINT(readability-function-cognitive-complexity)
+  LimitedSet<int, LimitedOptions<Size, Flags...>{}, Compare<int>> data;
   for (std::size_t len = 0; len < Size; ++len) {
     data.emplace(len * 100);
   }
+  std::size_t dropped = 0;
   while (!data.empty()) {
+    SCOPED_TRACE(absl::StrCat("Size: ", data.size()));
     for (std::size_t pos = 0; pos < Size + 1; ++pos) {
-      int v = 100 * static_cast<int>(pos + Size - data.size());
-      if (pos >= data.size()) {
+      const int v = static_cast<int>(100 * pos);
+      const std::size_t expected_pos = [&data, &v] {
+        for (std::size_t pos = 0; pos < data.size(); ++pos) {
+          if (data.at_index(pos) == v) {
+            return pos;
+          }
+        }
+        return data.npos;
+      }();
+      SCOPED_TRACE(absl::StrCat("Dropped: ", dropped, ", V: ", v, ", Expected: ", expected_pos));
+      if (expected_pos == data.npos) {
         ASSERT_THAT(data.index_of(v), data.npos);
         ASSERT_FALSE(data.contains(v));
         ASSERT_THAT(data.find(v), data.end());
       } else {
-        ASSERT_THAT(data.index_of(v), pos);
+        ASSERT_THAT(data.index_of(v), expected_pos);
         ASSERT_TRUE(data.contains(v));
         ASSERT_THAT(data.find(v), Ne(data.end()));
       }
     }
     data.erase(data.begin());
+    ++dropped;
   }
 }
 
-TEST_F(LimitedSetTest, CompareAllTheSizes) {
-  CompareAllTheSizes<1>();
-  CompareAllTheSizes<2>();
-  CompareAllTheSizes<3>();
-  CompareAllTheSizes<4>();
-  CompareAllTheSizes<5>();
-  CompareAllTheSizes<6>();
-  CompareAllTheSizes<7>();
-  CompareAllTheSizes<8>();
-  CompareAllTheSizes<9>();
-  CompareAllTheSizes<10>();
-  CompareAllTheSizes<11>();
-  CompareAllTheSizes<12>();
-  CompareAllTheSizes<13>();
-  CompareAllTheSizes<14>();
-  CompareAllTheSizes<15>();
-  CompareAllTheSizes<16>();
+template<template<typename> typename Compare, LimitedOptionsFlag Flags, std::size_t... Idx>
+void CompareAllTheSizes(const std::index_sequence<Idx...>& /*unused*/) {
+  (CompareAllTheSizesFor<Idx, Compare, Flags>(), ...);
 }
 
+template<template<typename> typename Compare, LimitedOptionsFlag Flags>
+void CompareAllTheSizes() {
+  CompareAllTheSizes<Compare, Flags>(std::make_index_sequence<50>());
+}
+
+// NOLINTBEGIN(google-readability-avoid-underscore-in-googletest-name)
+
+TEST_F(LimitedSetTest, CompareAllTheSizes_StdLess_Default) {
+  CompareAllTheSizes<std::less, LimitedOptionsFlag::kDefault>();
+}
+
+TEST_F(LimitedSetTest, CompareAllTheSizes_StdLess_NoOptimizeIndexOf) {
+  CompareAllTheSizes<std::less, LimitedOptionsFlag::kNoOptimizeIndexOf>();
+}
+
+TEST_F(LimitedSetTest, CompareAllTheSizes_CompareLess_Default) {
+  CompareAllTheSizes<mbo::types::CompareLess, LimitedOptionsFlag::kDefault>();
+}
+
+TEST_F(LimitedSetTest, CompareAllTheSizes_CompareLess_NoOptimizeIndexOf) {
+  CompareAllTheSizes<mbo::types::CompareLess, LimitedOptionsFlag::kNoOptimizeIndexOf>();
+}
+
+// NOLINTEND(google-readability-avoid-underscore-in-googletest-name)
 
 TEST_F(LimitedSetTest, PreSortedInput) {
   constexpr LimitedSet<int, LimitedOptions<4, LimitedOptionsFlag::kRequireSortedInput>{}> kData{0, 1, 2, 42};