+++ mbo/types/compare.h
@@ -15,11 +15,16 @@
 #ifndef MBO_TYPES_COMPARE_H_
 #define MBO_TYPES_COMPARE_H_
 
-#include <compare>// IWYU pragma: keep
+#include <compare>  // IWYU pragma: keep
+#include <functional>
 #include <type_traits>
 
 namespace mbo::types {
 
+// Compare less is a three-way comparator object that can be used as a drop-in
+// replacement for `std::less`. While it retains the `bool operator(l,r)` that
+// performs `less` comparison, it offers a complimentary `Compare` method which
+// performs three-way comparison in less order.
 template<std::three_way_comparable T>
 struct CompareLess {
   using value_type = T;  // NOLINT(readability-identifier-naming)
@@ -35,13 +40,13 @@ struct CompareLess {
 
   template<std::three_way_comparable_with<value_type> other_type>
   requires(!std::same_as<std::remove_cvref_t<value_type>, std::remove_cvref_t<other_type>>)
-  constexpr auto Compare(const value_type& lhs, other_type&& rhs) const noexcept {
+  constexpr auto Compare(const value_type& lhs, const other_type& rhs) const noexcept {
     return lhs <=> rhs;
   }
 
   template<std::three_way_comparable_with<value_type> other_type>
   requires(!std::same_as<std::remove_cvref_t<value_type>, std::remove_cvref_t<other_type>>)
-  constexpr auto Compare(other_type&& lhs, value_type&& rhs) const noexcept {
+  constexpr auto Compare(const other_type& lhs, const value_type& rhs) const noexcept {
     return lhs <=> rhs;
   }
 
@@ -49,13 +54,13 @@ struct CompareLess {
 
   template<std::three_way_comparable_with<value_type> other_type>
   requires(!std::same_as<std::remove_cvref_t<value_type>, std::remove_cvref_t<other_type>>)
-  constexpr bool operator()(const value_type& lhs, other_type&& rhs) const noexcept {
+  constexpr bool operator()(const value_type& lhs, const other_type& rhs) const noexcept {
     return (lhs <=> rhs) < 0;
   }
 
   template<std::three_way_comparable_with<value_type> other_type>
   requires(!std::same_as<std::remove_cvref_t<value_type>, std::remove_cvref_t<other_type>>)
-  constexpr bool operator()(other_type&& lhs, const value_type& rhs) const noexcept {
+  constexpr bool operator()(const other_type& lhs, const value_type& rhs) const noexcept {
     return (lhs <=> rhs) < 0;
   }
 };
@@ -68,10 +73,19 @@ struct IsCompareLess : std::false_type {};
 template<typename T>
 struct IsCompareLess<CompareLess<T>> : std::true_type {};
 
+template<typename T>
+struct IsStdLess : std::false_type {};
+
+template<typename T>
+struct IsStdLess<std::less<T>> : std::true_type {};
+
 }  // namespace types_internal
 
 template<typename T>
 concept IsCompareLess = types_internal::IsCompareLess<T>::value;
+
+template<typename T>
+concept IsStdLess = types_internal::IsStdLess<T>::value;
 
 }  // namespace mbo::types